Module iced_pure::widget

Use the built-in widgets or create your own.

Re-exports

pub use button::Button;

pub use container::Container;

pub use pane_grid::PaneGrid;

pub use pick_list::PickList;

pub use scrollable::Scrollable;

pub use slider::Slider;

pub use text_input::TextInput;

pub use tooltip::Tooltip;

pub use tree::Tree;

Modules

button

Allow your users to perform actions by pressing a button.

checkbox

Show toggle controls using checkboxes.

container

Decorate content and apply alignment.

image

Display images in your user interface.

pane_grid

Let your users split regions of your application and organize layout dynamically.

pick_list

Display a dropdown list of selectable values.

progress_bar

Provide progress feedback to your users.

radio

Create choices using radio buttons.

rule

Display a horizontal or vertical rule for dividing content.

scrollable

Navigate an endless amount of content with a scrollbar.

slider

Display an interactive selector of a single value from a range of values.

svg

Display vector graphics in your application.

text_input

Display fields that can be filled with text.

toggler

Show toggle controls using togglers.

tooltip

Display a widget over another.

tree

Store internal widget state in a state tree to ensure continuity.

Structs

Checkbox

A box that can be checked.

Column

A container that distributes its contents vertically.

Image

A frame that displays an image while keeping aspect ratio.

ProgressBar

A bar that displays progress.

Radio

A circular button representing a choice.

Row

A container that distributes its contents horizontally.

Rule

Display a horizontal or vertical rule for dividing content.

Space

An amount of empty space.

Svg

A vector graphics image.

Text

A paragraph of text.

Toggler

A toggler widget.

Enums

Position

The position of the tooltip. Defaults to following the cursor.

Traits

Widget

A component that displays information and allows interaction.