Struct iced_pure::widget::pick_list::PickList

pub struct PickList<'a, T, Message, Renderer> where
    [T]: ToOwned<Owned = Vec<T>>,
    Renderer: Renderer,
    Renderer::Theme: StyleSheet, { /* private fields */ }

A widget for selecting a single value from a list of options.

Implementations

impl<'a, T: 'a, Message, Renderer> PickList<'a, T, Message, Renderer> where
T: ToString + Eq,
[T]: ToOwned<Owned = Vec<T>>,
Renderer: Renderer,
Renderer::Theme: StyleSheet,

pub const DEFAULT_PADDING: Padding = _

The default padding of a PickList.

pub fn new(
options: impl Into<Cow<'a, [T]>>,
selected: Option<T>,
on_selected: impl Fn(T) -> Message + 'a
) -> Self

Creates a new PickList with the given list of options, the current selected value, and the message to produce when an option is selected.

pub fn placeholder(self, placeholder: impl Into<String>) -> Self

Sets the placeholder of the PickList.

pub fn width(self, width: Length) -> Self

Sets the width of the PickList.

pub fn padding<P: Into<Padding>>(self, padding: P) -> Self

Sets the Padding of the PickList.

pub fn text_size(self, size: u16) -> Self

Sets the text size of the PickList.

pub fn font(self, font: Renderer::Font) -> Self

Sets the font of the PickList.

pub fn style(
self,
style: impl Into<<Renderer::Theme as StyleSheet>::Style>
) -> Self

Sets the style of the PickList.