Struct iced_pure::widget::tooltip::Tooltip

pub struct Tooltip<'a, Message, Renderer: Renderer> { /* private fields */ }

An element to display a widget over another.

Implementations

impl<'a, Message, Renderer> Tooltip<'a, Message, Renderer> where
Renderer: Renderer,

pub fn new(
content: impl Into<Element<'a, Message, Renderer>>,
tooltip: impl ToString,
position: Position
) -> Self

Creates a new Tooltip.

pub fn size(self, size: u16) -> Self

Sets the size of the text of the Tooltip.

pub fn font(self, font: impl Into<Renderer::Font>) -> Self

Sets the font of the Tooltip.

pub fn gap(self, gap: u16) -> Self

Sets the gap between the content and its Tooltip.

pub fn padding(self, padding: u16) -> Self

Sets the padding of the Tooltip.

pub fn style(self, style_sheet: impl Into<Box<dyn StyleSheet + 'a>>) -> Self

Sets the style of the Tooltip.